+++ apps/prairielearn/src/ee/models/enrollment.sql
@@ -1,20 +1,3 @@
--- BLOCK insert_enrollment
-INSERT INTO
-  enrollments (user_id, course_instance_id)
-VALUES
-  ($user_id, $course_instance_id)
-RETURNING
-  *;
-
--- BLOCK select_enrollment_for_user_in_course_instance
-SELECT
-  *
-FROM
-  enrollments
-WHERE
-  user_id = $user_id
-  AND course_instance_id = $course_instance_id;
-
 -- BLOCK select_enrollment_counts
 SELECT
   (
